requirements.txt
streamlit
segno
pillow
reportlab
numpy
